+++ assetload.py
@@ -1,0 +1,82 @@
+import re
+
+quickidtable = ["NIC"]*102
+blockinfos = {}
+
+def frs(east, south, west, north, wt, flg):
+    wt = int(wt, 2)
+    info = int(east)*8+int(south)*4+int(west)*2+int(north)
+    print(info, east, south, west, north, flg, wt)
+    info = info & wt & flg 
+    print(str(bin(info))[2:].zfill(4))
+    info = list(str(bin(info))[2:].zfill(4))
+    print(flg, info)
+    east = int(east and info[0])*16
+    south = int(south and info[1])*16
+    west = int(west and info[2])*16
+    north = int(north and info[3])*16
+    return ((west  , north  , west+8  , north+8  ),
+            (east+8, north  , east+8+8, north+8  ),
+            (west  , south+8, west+8  , south+8+8),
+            (east+8, south+8, east+8+8, south+8+8))
+
+def getblockid():
+    with open("assets/block_id_.smp") as f:
+        pattern = re.compile(r"\{([a-zA-Z_]*)\}\s*:\s*\{([\d]*)\}")
+        for n, i in re.findall(pattern, f.read()):
+            try: blockinfos[n]
+            except: blockinfos[n] = {}
+            blockinfos[n]["id"] = int(i)
+            quickidtable[int(i)] = n
+
+def getblockpath():
+    with open("assets/block_textures.smp") as f:
+        pattern = re.compile(r"{(.*)}:{(.*)}")
+        for n, i in re.findall(pattern, f.read()):
+            try: blockinfos[n]
+            except: blockinfos[n] = {}
+            blockinfos[n]["path"] = i         
+
+def getblockcord():
+    with open("assets/block_icons.smp") as gbc:
+        pattern = re.compile(r"\{([a-zA-Z_]*)\}\s*:\s*\{\s*([\d]*),\s*([\d]*)\}")
+        for a, x, y in re.findall(pattern, gbc.read()):
+            blockinfos[a]["iconcord"] = (int(x), int(y))
+    return False
+
+def getlocal():
+    for fnm in "blocks credits hud input menu misc tutorial".split():
+        print(fnm)
+        with open("assets/localization/english_%s.txt" % fnm, "r") as f:
+            fc = re.sub(r"\\\s*\n", r"\\", f.read())
+            for line in fc.split("\n"):
+                for a, n, v in re.findall(r"^(\w*?) (\w*)\s*=\s*(.*)", line):
+                    try: blockinfos[n]
+                    except: blockinfos[n] = {}
+                    blockinfos[n][a] = v
+    for blkkey, item in blockinfos.items():
+        for blktype, txt in item.items():
+            for tartype, tarname in re.findall("{(\w+) (\w+)}", str(txt)):
+                if blockinfos[tarname]: 
+                    if blockinfos[tarname][tartype]:
+                        txt = re.sub("{%s %s}" % (tartype, tarname), blockinfos[tarname][tartype], txt)
+                        blockinfos[blkkey][blktype] = txt
+            for tartype, tarname, modifier in re.findall(r"{(\w+) (\w+)\|?([\^vsdbp]*)?}", str(txt)):
+                if blockinfos[tarname]:
+                    if blockinfos[tarname][tartype]:
+                        tx = blockinfos[tarname][tartype]
+                        for i in list(modifier):
+                            if   i == "^": tx = tx[0].upper()+tx[1:]
+                            elif i == "v": tx = tx[0].lower()+tx[1:]
+                            elif i == "s": tx = tx + "s"
+                            elif i == "d": tx = tx + "ed"
+                            elif i == "p": tx = tx + "'s"
+                            elif i == "b": tx = "{" + tx + "}"
+                            txt = txt.replace("{%s %s|%s}" % (tartype, tarname, modifier), tx)
+                        blockinfos[blkkey][blktype] = txt
+    return blockinfos
+def init():
+    getblockid()
+    getblockpath()
+    getblockcord()
+    getlocal()
