+++ main.py
@@ -1,5 +1,7 @@
 import os
+import glob
 import nextcord
+from assetload import init, blockinfos, quickidtable
 import keep_alive
 import datetime
 import random
@@ -15,63 +17,8 @@ intents.members = True
 intents.message_content = True
 bot = commands.Bot(command_prefix="!", intents=intents, help_command=None)
 print(cmd.ping.desc)
-quickidtable = ["NIC"]*102
-blockinfos = {}
-def getblockid():
-    with open("block_id_.smp") as f:
-        pattern = re.compile(r"\{([a-zA-Z_]*)\}\s*:\s*\{([\d]*)\}")
-        for n, i in re.findall(pattern, f.read()):
-            try: blockinfos[n]
-            except: blockinfos[n] = {}
-            blockinfos[n]["id"] = int(i)
-            quickidtable[int(i)] = n
+init()
 
-def getblockpath():
-    with open("block_textures.smp") as f:
-        pattern = re.compile(r"{(.*)}:{(.*)}")
-        for n, i in re.findall(pattern, f.read()):
-            try: blockinfos[n]
-            except: blockinfos[n] = {}
-            blockinfos[n]["path"] = i         
-
-def getblockcord():
-    with open("block_icons.smp") as gbc:
-        pattern = re.compile(r"\{([a-zA-Z_]*)\}\s*:\s*\{\s*([\d]*),\s*([\d]*)\}")
-        for a, x, y in re.findall(pattern, gbc.read()):
-            blockinfos[a]["iconcord"] = (int(x), int(y))
-    return False
-
-def getlocal():
-    for fnm in "blocks credits hud input menu misc tutorial".split():
-        print(fnm)
-        with open("localization/english_%s.txt" % fnm, "r") as f:
-            fc = re.sub(r"\\\s*\n", r"\\", f.read())
-            for line in fc.split("\n"):
-                for a, n, v in re.findall(r"^(\w*?) (\w*)\s*=\s*(.*)", line):
-                    try: blockinfos[n]
-                    except: blockinfos[n] = {}
-                    blockinfos[n][a] = v
-    for blkkey, item in blockinfos.items():
-        for blktype, txt in item.items():
-            for tartype, tarname in re.findall("{(\w+) (\w+)}", str(txt)):
-                if blockinfos[tarname]: 
-                    if blockinfos[tarname][tartype]:
-                        txt = re.sub("{%s %s}" % (tartype, tarname), blockinfos[tarname][tartype], txt)
-                        blockinfos[blkkey][blktype] = txt
-            for tartype, tarname, modifier in re.findall(r"{(\w+) (\w+)\|?([\^vsdbp]*)?}", str(txt)):
-                if blockinfos[tarname]:
-                    if blockinfos[tarname][tartype]:
-                        tx = blockinfos[tarname][tartype]
-                        for i in list(modifier):
-                            if   i == "^": tx = tx[0].upper()+tx[1:]
-                            elif i == "v": tx = tx[0].lower()+tx[1:]
-                            elif i == "s": tx = tx + "s"
-                            elif i == "d": tx = tx + "ed"
-                            elif i == "p": tx = tx + "'s"
-                            elif i == "b": tx = "{" + tx + "}"
-                            txt = txt.replace("{%s %s|%s}" % (tartype, tarname, modifier), tx)
-                        blockinfos[blkkey][blktype] = txt
-    return blockinfos
     
 @bot.command(name="help", description=cmd.help.desc, aliases=cmd.help.alias)
 async def help(ctx, tcmd=None,):
@@ -119,115 +66,6 @@ async def ping(ctx):
 async def scream(ctx, n:int=32):
     await ctx.send("A"*n)
    
-@bot.command(name="block", description=cmd.block.desc, aliases=cmd.block.alias)
-async def block(ctx, blk=None):
-    if blk:
-        for key, ite in blockinfos.items():
-            try: str(ite["id"])
-            except: break
-            val = str(ite["id"])
-            if blk == key or blk == val:
-                embed = nextcord.Embed()
-                
-                img = Image.open("block_zoo.png")
-                icox, icoy = blockinfos[key]["iconcord"]
-                img = img.crop((16*icox, 16*icoy, 16*(icox+1), 16*(icoy+1))).resize((128, 128), Image.NEAREST)
-                img.save("sed.png")
-                embed.title = ite["BLOCK_TITLE"]
-                embed.add_field(name="Block name", value=key)
-                embed.add_field(name="Block ID", value=val)
-                embed.add_field(name="Block Tutorial", value=re.sub(r"\\", r"\n", ite["BLOCK_TUTORIAL"]))
-                embed.set_image(url="attachment://sed.png")
-                
-                await ctx.send(file=nextcord.File("sed.png", filename="sed.png"), embed=embed)
-                return
-        await ctx.send(cmd.block.error % blk)
-    else:
-        await block(ctx, str(random.randint(0, 101)))
-
-def frs(east, south, west, north, wt, flg):
-    wt = int(wt, 2)
-    info = int(east)*8+int(south)*4+int(west)*2+int(north)
-    print(info, east, south, west, north, flg, wt)
-    info = info & wt & flg 
-    print(str(bin(info))[2:].zfill(4))
-    info = list(str(bin(info))[2:].zfill(4))
-    print(flg, info)
-    east = int(east and info[0])*16
-    south = int(south and info[1])*16
-    west = int(west and info[2])*16
-    north = int(north and info[3])*16
-    return ((west  , north  , west+8  , north+8  ),
-            (east+8, north  , east+8+8, north+8  ),
-            (west  , south+8, west+8  , south+8+8),
-            (east+8, south+8, east+8+8, south+8+8))
-    
-@bot.command(name="image", description=cmd.image.desc, aliases=cmd.image.alias)
-#eswn
-async def image(ctx, dt="[[16][20]][[16][16]]"):
-    blockp = []#3layer
-    cordic = []
-    cordwr = []
-    width, height = 0, 0 
-    dt.replace(" ", "")
-    dt = dt.lower()
-    for y, row in enumerate(re.findall(r"\[(\[.*?\])\]+", dt)):
-        for x, raw in enumerate(re.findall(r"\[([\w]+)#?([\d]{4}|)#?([\d]{1}|)\]", row)):
-            block, weldtag, rotation = raw
-            if not weldtag: weldtag = "1111"
-            if block.isdigit():
-                block = quickidtable[int(block)]
-                print(block, weldtag, rotation, x, y)
-            if block != "NIC" and block != "air":
-                if block in be.wiredtypes:
-                    if block in be.wafertypes:
-                        blockp += [("wafer", weldtag, x, y),
-                                    (block, "0000", x, y)]
-                        blockp += [("wire", weldtag, x, y)]
-                    if block in be.frametypes:
-                        blockp += [("frame", weldtag, x, y), 
-                                    (block, "0000", x, y)]
-                        blockp += [("wire", weldtag, x, y)]
-                    if block == "actuator":
-                        blockp += [("actuator_base", "0001", x, y), 
-                                ("actuator_head", "1111", x, y)]
-                    cordwr += [(x,y)]
-                    
-                else:
-                    blockp += [(block, weldtag, x, y)]
-                    print("T")
-            cordic += [(x, y)]
-        width = max(width, x+1)
-    height = max(height, y+1)
-    print("WIDTH, HEIGHT =", width, height)
-    fin = Image.new("RGBA", (width*16, height*16))
-    print(blockp)
-    for name, wt, x, y in blockp:#     e1,0   s0,1   w-1,0  n 0,-1 
-        print("")
-        print("name,wt,(x,y)", name, wt, (x, y))
-        print("BLOCKPATH = textures/blocks/"+blockinfos[name]["path"])
-        src = Image.open("textures/blocks/"+blockinfos[name]["path"]).convert("RGBA")
-        spflg = "1111"
-        findcord = cordwr if name == "wire" else cordic
-        for i, j in be.weldspr.items():
-            if name in j: 
-                spflg = i
-        
-        cord = frs(
-            (x+1, y) in findcord, 
-            (x, y+1) in findcord, 
-            (x-1, y) in findcord, 
-            (x, y-1) in findcord, 
-            wt,
-            int(spflg, 2))
-        print(cord, spflg)
-        for e, crd in enumerate(cord):
-            fin.alpha_composite(src.crop(crd), (x*16+(e%2)*8, y*16+(e//2)*8))
-    fin = fin.resize((width*16*2, height*16*2), Image.NEAREST)
-    fin.save("f.png")
-    await ctx.send(file=nextcord.File("f.png", filename="f.png"))
-            
-
 @bot.command(name="link", description=cmd.link.desc, aliases=cmd.link.alias)
 async def link(ctx, typ="r2d"):
     for i in keywords:
@@ -244,12 +82,12 @@ async def on_ready():
     global TimeOn
     TimeOn = datetime.datetime.now()
     
-    
-getblockid()
-getblockpath()
-getblockcord()
-getlocal()
-print(bot.all_commands["help"].description)
+# def returniscog(cogname):
+#     return bot.get_cog(cogname)
+
 token = os.environ['token']
+for cog_name in glob.glob("cog_*.py"):
+    print(cog_name, "LOAD")
+    bot.load_extension(cog_name[:-3])
 keep_alive.keep_alive()
 bot.run(token)
